+++ CandyBehavior.cs
@@ -19,12 +19,13 @@ namespace SCP956
     {
         private static ManualLogSource logger = Plugin.LoggerInstance;
 
-        public bool pinataCandy = true;
+        public AudioSource? ItemSFX; // TODO: Set audioclip to candycrunchsfx
 
         public override void Start()
         {
             base.Start();
-            // TODO: Set node sub text here?
+
+            ItemSFX.enabled = true;
         }
 
         public override void EquipItem()
@@ -46,7 +47,6 @@ namespace SCP956
             if (configEnableCandyBag.Value)
             {
                 logger.LogDebug("Putting candy in bag");
-                playerHeldBy.DespawnHeldObject();
 
                 if (playerHeldBy.ItemSlots.Where(x => x != null && x.itemProperties.name == "BagOfCandyItem").FirstOrDefault() == null)
                 {
@@ -57,7 +57,9 @@ namespace SCP956
                 if (candyBagObj == null) { logger.LogError("Candy bag not found"); return; }
 
                 CandyBagBehavior candyBag = candyBagObj.gameObject.GetComponent<CandyBagBehavior>();
-                candyBag.AddCandyToBagClientRpc(itemProperties.itemName, pinataCandy);
+                candyBag.AddCandyToBagClientRpc(itemProperties.itemName);
+
+                playerHeldBy.DespawnHeldObject();
             }
         }
 
@@ -81,100 +83,80 @@ namespace SCP956
             {
                 logger.LogDebug("Eating candy");
                 playerHeldBy.DespawnHeldObject();
-                ActivateCandy(itemProperties.itemName, pinataCandy);
+                ActivateCandy(itemProperties.itemName);
+                ItemSFX.Play();
             }
         }
 
-        public static void ActivateCandy(string candyName, bool pinataCandy)
+        public static void ActivateCandy(string candyName)
         {
-            HUDManager.Instance.UIAudio.PlayOneShot(CandyCrunchsfx, 1f);
+            logger.LogDebug("Activating candy effects...");
 
-            if (!pinataCandy || PlayerAge >= 12)
+            switch (candyName)
             {
-                if (pinataCandy && (int)UnityEngine.Random.Range(0, 101) < configCandyDeathChance.Value)
-                {
-                    localPlayer.KillPlayer(new Vector3(), true, CauseOfDeath.Unknown, 3);
-                    return;
-                }
-
-                if (!pinataCandy || configSecretLab.Value)
-                {
-                    logger.LogDebug("Activating candy effects...");
-
-                    switch (candyName)
+                case "Blue Candy":
+                    logger.LogDebug("Candy blue");
+                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyBlueEffects.Value); }
+                    else
                     {
-                        case "Blue Candy":
-                            logger.LogDebug("Candy blue");
-                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyBlueEffects.Value); }
-                            else
-                            {
-                                StatusEffectController.Instance.HealPlayer(30, true);
-                            }
-                            break;
-                        case "Green Candy":
-                            logger.LogDebug("Candy green");
-                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyGreenEffects.Value); }
-                            else
-                            {
-                                StatusEffectController.Instance.StatusNegation(30);
-                                StatusEffectController.Instance.HealthRegen(1, 80);
-                            }
-                            break;
-                        case "Purple Candy":
-                            logger.LogDebug("Candy purple");
-                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyPurpleEffects.Value); }
-                            else
-                            {
-                                StatusEffectController.Instance.DamageReduction(15, 20, true);
-                                StatusEffectController.Instance.HealthRegen(2, 10);
-                            }
-                            break;
-                        case "Red Candy":
-                            logger.LogDebug("Candy red");
-                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyRedEffects.Value); }
-                            else
-                            {
-                                StatusEffectController.Instance.HealthRegen(9, 5);
-                            }
-                            break;
-                        case "Yellow Candy":
-                            logger.LogDebug("Candy yellow");
-                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyYellowEffects.Value); }
-                            else
-                            {
-                                StatusEffectController.Instance.RestoreStamina(25);
-                                StatusEffectController.Instance.InfiniteSprint(8);
-                                StatusEffectController.Instance.IncreasedMovementSpeed(8, 2, true, true);
-                            }
-                            break;
-                        case "Pink Candy":
-                            logger.LogDebug("Candy pink");
-                            Landmine.SpawnExplosion(localPlayer.transform.position, true, 3, 3);
-                            break;
-                        case "Rainbow Candy":
-                            logger.LogDebug("Candy rainbow");
-                            StatusEffectController.Instance.HealPlayer(15);
-                            StatusEffectController.Instance.InfiniteSprint(5, true);
-                            StatusEffectController.Instance.bulletProofMultiplier += 1;
-                            StatusEffectController.Instance.StatusNegation(10);
-                            StatusEffectController.Instance.HealPlayer(20, true);
-                            break;
-                        case "Black Candy":
-                            logger.LogDebug("Candy black");
-                            ActivateCandy(CandyNames.Where(x => x != "Black Candy").ToList()[UnityEngine.Random.Range(0, CandyNames.Count - 1)], pinataCandy);
-                            break;
-                        default:
-                            logger.LogDebug("Candy not found");
-                            break;
+                        StatusEffectController.Instance.HealPlayer(30, true);
                     }
-                }
-            }
-            else if (configEnablePinata.Value)
-            {
-                // TODO: Animation for player turning into SCP956 and bones crunching sound effects. Maybe spawn in as scavenger model and play animation to turn into scp956!
-                localPlayer.KillPlayer(new Vector3(), false, CauseOfDeath.Unknown, 3);
-                int index = RoundManager.Instance.currentLevel.Enemies.FindIndex(x => x.enemyType.enemyName == "SCP-956");
-                RoundManager.Instance.SpawnEnemyOnServer(localPlayer.transform.position, localPlayer.previousYRot, index);
+                    break;
+                case "Green Candy":
+                    logger.LogDebug("Candy green");
+                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyGreenEffects.Value); }
+                    else
+                    {
+                        StatusEffectController.Instance.StatusNegation(30);
+                        StatusEffectController.Instance.HealthRegen(1, 80);
+                    }
+                    break;
+                case "Purple Candy":
+                    logger.LogDebug("Candy purple");
+                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyPurpleEffects.Value); }
+                    else
+                    {
+                        StatusEffectController.Instance.DamageReduction(15, 20, true);
+                        StatusEffectController.Instance.HealthRegen(2, 10);
+                    }
+                    break;
+                case "Red Candy":
+                    logger.LogDebug("Candy red");
+                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyRedEffects.Value); }
+                    else
+                    {
+                        StatusEffectController.Instance.HealthRegen(9, 5);
+                    }
+                    break;
+                case "Yellow Candy":
+                    logger.LogDebug("Candy yellow");
+                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyYellowEffects.Value); }
+                    else
+                    {
+                        StatusEffectController.Instance.RestoreStamina(25);
+                        StatusEffectController.Instance.InfiniteSprint(8);
+                        StatusEffectController.Instance.IncreasedMovementSpeed(8, 2, true, true);
+                    }
+                    break;
+                case "Pink Candy":
+                    logger.LogDebug("Candy pink");
+                    Landmine.SpawnExplosion(localPlayer.transform.position, true, 3, 3);
+                    break;
+                case "Rainbow Candy":
+                    logger.LogDebug("Candy rainbow");
+                    StatusEffectController.Instance.HealPlayer(15);
+                    StatusEffectController.Instance.InfiniteSprint(5, true);
+                    StatusEffectController.Instance.bulletProofMultiplier += 1;
+                    StatusEffectController.Instance.StatusNegation(10);
+                    StatusEffectController.Instance.HealPlayer(20, true);
+                    break;
+                case "Black Candy":
+                    logger.LogDebug("Candy black");
+                    ActivateCandy(CandyNames.Where(x => x != "Black Candy").ToList()[UnityEngine.Random.Range(0, CandyNames.Count - 1)]);
+                    break;
+                default:
+                    logger.LogDebug("Candy not found");
+                    break;
             }
         }
     }